+++ kraken/lib/src/dom/elements/input.dart
@@ -57,7 +57,7 @@ const Duration _kCursorBlinkWaitForStart = Duration(milliseconds: 150);
 const TextSelection blurSelection = TextSelection.collapsed(offset: -1);
 
 class EditableTextDelegate implements TextSelectionDelegate {
-  InputElement _inputElement;
+  final InputElement _inputElement;
   EditableTextDelegate(this._inputElement);
 
   TextEditingValue _textEditingValue = TextEditingValue();
@@ -119,7 +119,7 @@ class InputElement extends Element implements TextInputClient, TickerProvider {
     }
   }
 
-  static SplayTreeMap<int, InputElement> _nativeMap = SplayTreeMap();
+  static final SplayTreeMap<int, InputElement> _nativeMap = SplayTreeMap();
 
   static InputElement getInputElementOfNativePtr(Pointer<NativeInputElement> nativePtr) {
     InputElement? element = _nativeMap[nativePtr.address];
@@ -285,7 +285,7 @@ class InputElement extends Element implements TextInputClient, TickerProvider {
     _textSelectionDelegate.userUpdateTextEditingValue(value, SelectionChangedCause.keyboard);
     TextSpan? text = obscureText ? _buildPasswordTextSpan(_actualText!.text!) : _actualText;
     if (_renderEditable != null) {
-      _renderEditable!.text = _actualText!.text!.length == 0
+      _renderEditable!.text = _actualText!.text!.isEmpty
           ? placeholderTextSpan
           : text;
     }
@@ -395,11 +395,9 @@ class InputElement extends Element implements TextInputClient, TickerProvider {
   int _maxLength = 9007199254740992;
 
   RenderEditable createRenderEditable() {
-    if (_actualText == null) {
-      _actualText = _buildTextSpan();
-    }
+    _actualText ??= _buildTextSpan();
     TextSpan text = _actualText!;
-    if (_actualText!.toPlainText().length == 0) {
+    if (_actualText!.toPlainText().isEmpty) {
       text = placeholderTextSpan;
     } else if (obscureText) {
       text = _buildPasswordTextSpan(text.text!);
@@ -449,7 +447,7 @@ class InputElement extends Element implements TextInputClient, TickerProvider {
     switch (action) {
       case TextInputAction.done:
         _triggerChangeEvent();
-        blur();
+        InputElement.clearFocus();
         break;
       case TextInputAction.none:
         // TODO: Handle this case.
@@ -573,7 +571,7 @@ class InputElement extends Element implements TextInputClient, TickerProvider {
 
   void _handleTextChanged(String text, bool userInteraction, SelectionChangedCause? cause) {
     if (_renderEditable != null) {
-      if (text.length == 0) {
+      if (text.isEmpty) {
         _renderEditable!.text = placeholderTextSpan;
       } else if (obscureText) {
         _renderEditable!.text = _buildPasswordTextSpan(text);
@@ -877,6 +875,7 @@ class InputElement extends Element implements TextInputClient, TickerProvider {
     print('ShowAutocorrectionPromptRect start: $start, end: $end');
   }
 
+  @override
   void dispose() {
     super.dispose();
     _nativeMap.remove(nativeInputElement.address);