+++ kraken/lib/src/module/method_channel.dart
@@ -83,6 +83,7 @@ class KrakenJavaScriptChannel extends KrakenMethodChannel {
     _methodCallCallback = value;
   }
 
+  @override
   Future<dynamic> _invokeMethodFromJavaScript(String method, List arguments) {
     if (_methodCallCallback == null) return Future.value(null);
     return _methodCallCallback!(method, arguments);
@@ -92,7 +93,8 @@ class KrakenJavaScriptChannel extends KrakenMethodChannel {
 class KrakenNativeChannel extends KrakenMethodChannel {
   // Flutter method channel used to communicate with public SDK API
   // Only works when integration wieh public SDK API
-  static MethodChannel _nativeChannel = getKrakenMethodChannel()
+
+  static final MethodChannel _nativeChannel = getKrakenMethodChannel()
     ..setMethodCallHandler((call) async {
       String method = call.method;
       KrakenController? controller = KrakenController.getControllerOfJSContextId(0);
@@ -108,6 +110,7 @@ class KrakenNativeChannel extends KrakenMethodChannel {
       return Future<dynamic>.value(null);
     });
 
+  @override
   Future<dynamic> _invokeMethodFromJavaScript(String method, List arguments) async {
     Map<String, dynamic> argsWrap = {
       'method': method,